Split Interpreter from its result
